// src/generated/prisma/models/AssignedTicket.ts
export type AssignedTicketUncheckedCreateInput = {
  techId: string
  ticketId: string
  status?: $Enums.Status
  completed?: boolean
  assignBy: string
  assignAt?: Date | string
}
// ...
export type AssignedTicketUncheckedUpdateInput = {
  techId?: Prisma.StringFieldUpdateOperationsInput | string
  ticketId?: Prisma.StringFieldUpdateOperationsInput | string
  status?: Prisma.EnumStatusFieldUpdateOperationsInput | $Enums.Status
  completed?: Prisma.BoolFieldUpdateOperationsInput | boolean
  assignBy?: Prisma.StringFieldUpdateOperationsInput | string
  assignAt?: Prisma.DateTimeFieldUpdateOperationsInput | Date | string
}
// ...
export type EnumStatusFieldUpdateOperationsInput = {
  set?: $Enums.Status
}

export type BoolFieldUpdateOperationsInput = {
  set?: boolean
}

export type StringFieldUpdateOperationsInput = {
  set?: string
}

export type DateTimeFieldUpdateOperationsInput = {
  set?: Date | string
}